fix print formatting

+++ Core/Src/drone_can/drone_can.hpp
@@ -5,6 +5,6 @@
 
 namespace DroneCan {
 	// return 0 for OK, negative for error.
-  int init(CAN_HandleTypeDef *hcan, const uint8_t node_id);
+  int16_t init(CAN_HandleTypeDef *hcan, const uint8_t node_id);
   void update();
 }
